requirements.txt
# Python 3.6.6
# code format
flake8==3.7.7

# Bleeding edge Flask
Flask==1.1.1

# Your custom requirements go here

# database
PyMySQL==0.9.2

# Sentry
raven==6.9.0

# http
requests==2.19.1

# Cache
Flask-Redis==0.3.0

# Queue
Flask-Celery==2.4.3

# migration
alembic==1.3.1
Flask-Migrate==2.2.1

# orm
Flask-SQLAlchemy==2.3.2

# script
Flask-Script==2.0.6

# signal
blinker==1.4

# WSGI Handler
# ------------------------------------------------
gevent==1.4.0
gunicorn==19.7.1

# form validate
marshmallow==3.5.1

# api doc
flask_restplus==0.12.1

# uuid
shortuuid==0.5.0

# werkzeug  import bug
werkzeug==0.16.0

# PIL
Pillow==4.1.0

# captcha
captcha==0.3
